% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_dapps.r
\name{dr_dapps}
\alias{dr_dapps}
\title{Get Multiple Dapps}
\usage{
dr_dapps(
  base_url_token,
  chain = NULL,
  page = NULL,
  resultsPerPage = NULL,
  api_key
)
}
\arguments{
\item{base_url_token}{Token included into Base URL}

\item{chain}{Chain to search for dapps}

\item{page}{Page of search results}

\item{resultsPerPage}{Expected results per page}

\item{api_key}{API key}
}
\description{
Get Multiple Dapps <https://api-portal.dappradar.com/catalog/prd_p77wm84fbe4gteuu/endpoint/fltr_K56Bb9ncytxQO1X2egHdDn5ljwxhljng>
}
